requirements.txt
# Core Dependencies
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Visualization
matplotlib>=3.7.0
plotly>=5.14.0
pyvista>=0.43.0
vtk>=9.2.0

# Web Framework
flask>=2.3.0
flask-socketio>=5.3.0

# Data Processing
scikit-learn>=1.3.0
statsmodels>=0.14.0

# 3D Model Processing
trimesh>=3.21.0
pygltflib>=1.16.0

# Reporting
reportlab>=4.0.0
jinja2>=3.1.0
python-docx>=0.8.11

# Database
sqlalchemy>=2.0.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
